requirements.txt
torch==2.1.0+cu121
numpy==1.24.3
PyYAML==6.0.1
requests==2.31.0
pandas==2.1.3
python-dotenv==1.0.0